services_router-webhook/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
google-cloud-logging==3.*   # ← add this line
pandas==2.*
